pyproject.toml
[tool.poetry]
name = "exportar documentos"
version = "0.1"
description = ""
authors = ["Eliezer"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.3"
pika = "^1.3.1"
gunicorn = "^20.1.0"
inject = "^4.3.1"
flask-sqlalchemy = "^3.0.3"
google-cloud-storage = "^2.8.0"
redis = "^4.5.4"


[tool.poetry.group.dev.dependencies]
poethepoet = "^0.19.0"
autopep8 = "^2.0.2"
pytest = "^7.3.1"
pytest-mock = "^3.10.0"
coverage = "^7.2.4"

[build-system]
requires = ["poetry-core=1.4.2"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks.start]
cmd = "flask run --port 8080"
help = "Starts the application"

[tool.poe.tasks.cove_tests]
cmd = "poetry run coverage run -m pytest"
help = "Run tests your project"

[tool.poe.tasks.cove_report]
cmd = "poetry run coverage report"
help = "Reports on the test coverage"

[tool.poe.tasks.cove_html]
cmd = "poetry run coverage html"
help = "Generates HTML reports on the test coverage"
